fix/ltx2-vocoder-bf16-dtype-mismatch

+++ python/sglang/multimodal_gen/runtime/models/vocoder/ltx_2_vocoder.py
@@ -707,7 +707,7 @@ class LTX2Vocoder(ABC, nn.Module, LayerwiseOffloadableModuleMixin):
                 else nullcontext()
             )
             with autocast_ctx:
-                waveform = self.vocoder(hidden_states.float())
+                waveform = self.vocoder(hidden_states)
                 length_low_rate = waveform.shape[-1]
                 output_length = (
                     length_low_rate
